Split out unit tests into separate files
